refactoring code post and update

+++ back/controllers/postCtrl.js
@@ -1,11 +1,19 @@
 import Post from '../models/post.js';
 import User from '../models/user.js';
+import fs from "fs"
 
 // affichage de tous les posts
 const getAll = (req, res, next) => {
     Post.findAll({include:[User]})
     .then(posts => res.status(200).json(posts))
     .catch(error => res.status(500).json({msg: ""+error}))
+}
+
+const getById = (req, res) =>{
+    let id = req.params.id;
+    Post.findByPk(id)
+    .then(post => res.status(200).json(post))
+    .catch(error => res.status(500).json({msg : "" + error}))
 }
 
 // création d'un post
@@ -27,10 +35,18 @@ const createPost = (req, res, next) => {
 const updatePost = async (req, res, next) => {
     const {id} = req.params;
     const {body} = req;
+    // let post =JSON.parse(req.body.post);
+    // let image = req.file;
+    // if(image){
+    //     post = {
+    //         ...post,
+    //         image: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
+    //     }
+    // }
 
     try {
         //recherche de l'identifiant du post
-        let post = await  Post.findByPk(id)
+        let post = await Post.findByPk(id)
         if(!post)return res.status(404).json({msg : "post not found"});
         post.content = body.content;
         post.commentaire = body.commentaire;
@@ -45,14 +61,25 @@ const updatePost = async (req, res, next) => {
 // supression d'un post
 const deletePost = async (req, res) => {
     const {id} =req.params;
-
     try {
-        let ressource = await Post.destroy({where : {id : id}})
-        if (ressource === 0) return res.status(404).json({msg: "Not found"})
-        res.status(200).json({msg: "Deleted post"})
+        let post = await Post.findByPk(id)
+        if(!post) return res.status(404).json({msg : "post not found"});
+        if(post.image === null){
+            let ressource = Post.destroy({where : {id : id}})
+            if (ressource === 0) return res.status(404).json({msg: "Not found"})
+            res.status(200).json({msg: "Deleted post"})
+        }else{
+            const filename = post.image.split('/images/')[1];
+            console.log(filename);
+            fs.unlink(`images/${filename}`, () => {
+                let ressource = Post.destroy({where : {id : id}})
+                if (ressource === 0) return res.status(404).json({msg: "Not found"})
+                res.status(200).json({msg: "Deleted post"})
+            })
+        }
     } catch (error) {
         return res.status(500).json({msg : "Database Error", error : error})
     }
 }
 
-export {getAll, createPost, updatePost, deletePost};
+export {getAll, createPost, updatePost, deletePost, getById};
